Beginning of Bonus part 2

+++ push_swap/libft/ft_atoi.c
@@ -12,12 +12,12 @@
 
 #include "libft.h"
 
-int	ft_atoi(const char *nptr)
+long int	ft_atoi(const char *nptr)
 {
-	int	i;
-	int	c;
-	int	res;
-	int	n;
+	int		i;
+	int		c;
+	long int	res;
+	int		n;
 
 	i = 0;
 	c = 1;
